Attempts to draw strokes to image, with no effect

+++ src/parse.rs
@@ -16,7 +16,7 @@ pub enum Node {
 
 fn parse_vertices(data: String) -> Vec<Point> {
     data.split(',').filter_map(|pos| {
-        let xy = pos.split(' ').filter_map(|s| s.parse::<f64>().ok()).collect::<Vec<_>>();
+        let xy = pos.split(' ').filter_map(|s| s.parse::<f32>().ok()).collect::<Vec<_>>();
         if xy.len() > 0 {
             Some(Point(xy[0].clone(), xy[1].clone() ))
         } else {
